+++ saleor/graphql/api.py
@@ -15,6 +15,7 @@ from .shipping.schema import ShippingMutations, ShippingQueries
 from .shop.schema import ShopMutations, ShopQueries
 from .translations.schema import TranslationQueries
 from .webhook.schema import WebhookMutations, WebhookQueries
+from .wishlist.schema import WishlistMutations
 
 
 class Query(
@@ -52,6 +53,7 @@ class Mutation(
     ShippingMutations,
     ShopMutations,
     WebhookMutations,
+    WishlistMutations,
 ):
     pass