+++ saleor/graphql/core/types/common.py
@@ -15,6 +15,7 @@ from ..enums import (
     ShippingErrorCode,
     ShopErrorCode,
     WebhookErrorCode,
+    WishlistErrorCode,
 )
 from .money import VAT
 
@@ -87,6 +88,10 @@ class ExtensionsError(Error):
 
 class WebhookError(Error):
     code = WebhookErrorCode(description="The error code.")
+
+
+class WishlistError(Error):
+    code = WishlistErrorCode(description="The error code.")
 
 
 class LanguageDisplay(graphene.ObjectType):