+++ CHANGELOG.md
@@ -34,6 +34,7 @@ All notable, unreleased changes to this project will be documented in this file.
 - Drop storefront 1.0 - #5043 by @IKarbowiak
 - Replace permissions strings with enums - #5038 by @kswiatek92
 - Remove gateways forms and templates - #5075 by @IKarbowiak
+- Add `Wishlist` models and GraphQL endpoints - #5021 by @derenio
 
 ## 2.9.0